SONARKT-235 Improve spotless configuration *Update spotless to the latest version 6.11.0 *configure spotless for all modules except kotlin-checks-test-sources
